refactor: move database stuff into room package
