+++ cardinal-android/app/src/main/java/earth/maps/cardinal/routing/FerrostarWrapperRepository.kt
@@ -21,6 +21,7 @@ import com.stadiamaps.ferrostar.core.AndroidTtsObserver
 import dagger.hilt.android.qualifiers.ApplicationContext
 import earth.maps.cardinal.data.AppPreferenceRepository
 import earth.maps.cardinal.data.RoutingMode
+import earth.maps.cardinal.data.room.RoutingProfileRepository
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -29,7 +30,7 @@ class FerrostarWrapperRepository @Inject constructor(
     @param:ApplicationContext
     private val context: Context,
     private val appPreferences: AppPreferenceRepository,
-    private val routingProfileRepository: earth.maps.cardinal.data.RoutingProfileRepository
+    private val routingProfileRepository: RoutingProfileRepository
 ) {
     lateinit var walking: FerrostarWrapper
     lateinit var cycling: FerrostarWrapper
