+++ cardinal-android/app/src/main/java/earth/maps/cardinal/tileserver/TileDownloadForegroundService.kt
@@ -34,10 +34,10 @@ import androidx.core.app.NotificationCompat
 import dagger.hilt.android.AndroidEntryPoint
 import earth.maps.cardinal.MainActivity
 import earth.maps.cardinal.R
-import earth.maps.cardinal.data.DownloadStatus
-import earth.maps.cardinal.data.DownloadedTileDao
-import earth.maps.cardinal.data.OfflineAreaDao
-import earth.maps.cardinal.data.TileType
+import earth.maps.cardinal.data.room.DownloadStatus
+import earth.maps.cardinal.data.room.DownloadedTileDao
+import earth.maps.cardinal.data.room.OfflineAreaDao
+import earth.maps.cardinal.data.room.TileType
 import earth.maps.cardinal.geocoding.TileProcessor
 import earth.maps.cardinal.ui.BoundingBox
 import kotlinx.coroutines.CoroutineScope
