+++ cardinal-android/app/src/main/java/earth/maps/cardinal/tileserver/TileDownloadManager.kt
@@ -25,12 +25,12 @@ import android.database.sqlite.SQLiteDatabase
 import android.os.IBinder
 import android.util.Log
 import earth.maps.cardinal.R
-import earth.maps.cardinal.data.DownloadStatus
-import earth.maps.cardinal.data.DownloadedTile
-import earth.maps.cardinal.data.DownloadedTileDao
-import earth.maps.cardinal.data.OfflineArea
-import earth.maps.cardinal.data.OfflineAreaDao
-import earth.maps.cardinal.data.TileType
+import earth.maps.cardinal.data.room.DownloadStatus
+import earth.maps.cardinal.data.room.DownloadedTile
+import earth.maps.cardinal.data.room.DownloadedTileDao
+import earth.maps.cardinal.data.room.OfflineArea
+import earth.maps.cardinal.data.room.OfflineAreaDao
+import earth.maps.cardinal.data.room.TileType
 import earth.maps.cardinal.geocoding.TileProcessor
 import earth.maps.cardinal.ui.BoundingBox
 import io.ktor.client.HttpClient
