+++ cardinal-android/app/src/main/java/earth/maps/cardinal/ui/AppContent.kt
@@ -77,7 +77,7 @@ import earth.maps.cardinal.R.dimen
 import earth.maps.cardinal.R.drawable
 import earth.maps.cardinal.data.AppPreferenceRepository
 import earth.maps.cardinal.data.LatLng
-import earth.maps.cardinal.data.OfflineArea
+import earth.maps.cardinal.data.room.OfflineArea
 import earth.maps.cardinal.data.Place
 import earth.maps.cardinal.ui.settings.AccessibilitySettingsScreen
 import earth.maps.cardinal.ui.settings.AdvancedSettingsScreen
