+++ cardinal-android/app/src/main/java/earth/maps/cardinal/viewmodel/OfflineAreasViewModel.kt
@@ -29,8 +29,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
-import earth.maps.cardinal.data.OfflineArea
-import earth.maps.cardinal.data.OfflineAreaRepository
+import earth.maps.cardinal.data.room.OfflineArea
+import earth.maps.cardinal.data.room.OfflineAreaRepository
 import earth.maps.cardinal.tileserver.TileDownloadForegroundService
 import earth.maps.cardinal.tileserver.calculateTileRange
 import earth.maps.cardinal.ui.BoundingBox
