+++ cardinal-android/app/src/main/java/earth/maps/cardinal/viewmodel/PlaceCardViewModel.kt
@@ -21,8 +21,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import earth.maps.cardinal.data.Place
-import earth.maps.cardinal.data.PlaceDao
-import earth.maps.cardinal.data.PlaceEntity
+import earth.maps.cardinal.data.room.PlaceDao
+import earth.maps.cardinal.data.room.PlaceEntity
 import kotlinx.coroutines.launch
 import javax.inject.Inject
